requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==24.2.0
bip-utils==2.9.3
cbor2==5.6.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
coincurve==20.0.0
colorama==0.4.6
crcmod==1.7
ecdsa==0.19.0
ed25519-blake2b==1.4.1
fastapi==0.115.12
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
grequests==0.7.0
h11==0.16.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
pyngrok==7.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.46.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.2
yarl==1.17.1
zope.event==5.0
zope.interface==7.1.1
sqlcipher3-wheels
flask
replit-object-storage
